the "Uncaught ChunkLoadError" issue by modifying `public/sw.js`. solved
The service worker now includes a `controllerchange` event listener. When a new version of the service worker takes control of the page, it will automatically trigger a full page reload. This ensures that the browser fetches the latest HTML and all corresponding JavaScript chunks, preventing inconsistencies that lead to `ChunkLoadError` after deployments.